fix master

+++ backend-pegasus/api/routes/student/deletestudentcertificate.js
@@ -1,7 +1,12 @@
 const {env, sha1, mysql, mypool} = require('../../util')
 
 const deletestudentcertificate = (req, res) => {
-    const studentid = req.params.StudentID;
+    const studentid = req.body.StudentID;
+    const certid = req.body.CertificateID;
+    const certname = req.body.Name;
+    const issuedby = req.body.IssuedBy;            
+    var issueddate = req.body.IssueDate;
+    var validuntil = req.body.ValidUntil;
 
     mypool.getConnection( (error, connection) => {
         if(error) {
@@ -11,7 +16,10 @@ const deletestudentcertificate = (req, res) => {
         }
         else {
             if(studentid) {   
-                let queryString1 = `delete from pegasus.studentcertificate where studentid = "${studentid}"` ;  
+                let queryString1 = `delete from pegasus.studentcertificate where id = "${certid}" and studentid = "${studentid}"` ;   
+                if(!studentid) {
+                    queryString1 = `delete from pegasus.studentcertificate where studentid = "${studentid}"` ;   
+                }       
                 connection.query(queryString1, (err, rows, fields) => {
                     if(err) {
                         res.status(500).json({ message: err })
